Are two words anagrams?
In a class Anagrams, implement a static method boolen areAnagrams(String a, String b).

For anagrams, check here - https://en.wikipedia.org/wiki/Anagram

For example, listen and silent are anagrams.

The program should read two words from the standard input and output:

ANAGRAMS if the words are anagrams of each other
NOT ANAGRAMS if the two words are not anagrams of each other
Consider lowering the case of the two words since the case does not matter. SILENT and listen are also anagrams.